+++ src/dyscord/objects/interactions/__init__.py
@@ -4,7 +4,7 @@ For more information, see [Discord API](https://discord.com/developers/docs/inte
 '''
 
 from .command import Command, CommandOptions, CommandOptionChoiceStructure
-from .interaction import InteractionStructure, InteractionResponse
+from .interaction import Interaction, InteractionResponse
 from .components import Component, ActionRow, Button, SelectMenu
 from .enumerations import CHANNEL_TYPE, COMMAND_OPTION, COMMAND_TYPE, BUTTON_STYLES, INTERACTION_TYPES, INTERACTION_RESPONSE_TYPES,\
     INTERACTION_CALLBACK_FLAGS
@@ -13,7 +13,7 @@ __all__ = [
     'Command',
     'CommandOptions',
     'CommandOptionChoiceStructure',
-    'InteractionStructure',
+    'Interaction',
     'CHANNEL_TYPE',
     'COMMAND_OPTION',
     'COMMAND_TYPE',