+++ docs/changelog.md
@@ -2,13 +2,17 @@
 
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.3.2] WIP
+## [v0.4.0] WIP
 ### Added
 - New set_all_intents() function in DiscordClient. Sets all intents to True.
 - Better example code.
 - User and Member now have all attributes set to `None` by default.
 - Role now has all attributes set to `None` by default.
 - Begin restructure of documentation to make some sense.
+
+### Changed
+- `InteractionStructure` renamed to `Interaction`.
+- `InteractionDataStructure` renamed to `InteractionData`.
 
 ### Fixed
 - Exposed the `__version__` string at the top module level.